requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
decorator==4.4.2
Flask==1.1.2
Flask-Cors==3.0.9
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-dotenv==0.15.0
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.26.2
validators==0.18.1
Werkzeug==1.0.1
